comments resolved
